+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 v3.8
 
 - Fixed content menu not hiding mod menu.
+- [Spectate] Fixed second spectate button appearing on 1v1 games.
 - [UIURLS] Added support for new content menu.
 
 v3.7